Refactor code
- Add ESLint & Prettier
- Move files so that the codes are more understandable
