+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules
 package-lock.json
 todos.json
+dist
+yarn.lock
