+++ .prettierrc
@@ -1,0 +1,6 @@
+{
+	"semi": false,
+	"singleQuote": true,
+	"printWidth": 100,
+	"useTabs": true
+}
