+++ package.json
@@ -1,18 +1,30 @@
 {
-  "name": "to-do-app",
-  "version": "1.0.0",
-  "description": "",
-  "main": "src/main.js",
-  "scripts": {
-    "start": "electron ."
-  },
-  "author": "boraoksuzoglu",
-  "license": "MIT",
-  "devDependencies": {
-    "electron": "^20.0.2"
-  },
-  "dependencies": {
-    "@electron/remote": "^2.0.8",
-    "fs": "^0.0.1-security"
-  }
+	"name": "to-do-app",
+	"version": "1.1.0",
+	"description": "",
+	"main": "src/main.js",
+	"scripts": {
+		"pack": "electron-builder --dir",
+		"dist": "electron-builder",
+		"start": "electron ."
+	},
+	"author": "boraoksuzoglu",
+	"license": "MIT",
+	"devDependencies": {
+		"electron": "^20.0.2",
+		"electron-builder": "^23.3.3",
+		"eslint": "^8.22.0",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-plugin-jest": "^26.8.4",
+		"eslint-plugin-prettier": "^4.2.1",
+		"prettier": "^2.7.1"
+	},
+	"dependencies": {
+		"@electron/remote": "^2.0.8",
+		"fs": "^0.0.1-security",
+		"lodash": "^4.17.21"
+	},
+	"build": {
+		"target": "deb"
+	}
 }
