+++ src/renderer/assets/styles.css
@@ -3,11 +3,11 @@ html {
 }
 
 body {
-	background: #FF416C;
+	background: #ff416c;
 	/* fallback for old browsers */
-	background: -webkit-linear-gradient(to left, #FF4B2B, #FF416C);
+	background: -webkit-linear-gradient(to left, #ff4b2b, #ff416c);
 	/* Chrome 10-25, Safari 5.1-6 */
-	background: linear-gradient(to left, #FF4B2B, #FF416C);
+	background: linear-gradient(to left, #ff4b2b, #ff416c);
 	/* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
 
 	background-repeat: no-repeat;
@@ -35,11 +35,11 @@ input {
 	overflow: auto;
 }
 
-.column>* {
+.column > * {
 	margin-bottom: 0.3rem;
 }
 
-.column> :last-child {
+.column > :last-child {
 	margin-bottom: 0;
 }
